+++ ios/Classes/LlamaDartPlugin.swift
@@ -1,6 +1,6 @@
 //
 //  LlamaDartPlugin.swift
-//  llama_dart
+//  llamadart
 //
 //  This is a dummy plugin class to satisfy CocoaPods requirements.
 //  The actual functionality is implemented via FFI bindings.